refactor(repository): improve in-memory msg repo

+++ internal/interfaces/message-repository.go
@@ -4,6 +4,6 @@ import "github.com/ivankatalenic/web-chat/internal/models"
 
 // MessageRepository is a message repository
 type MessageRepository interface {
-	GetLast(n int) (models.MessageSlice, error)
+	GetLast(n int64) (models.MessageSlice, error)
 	Put(msg *models.Message) error
 }
